move external tool related types from store to data module and merge some type files
