+++ src/modules/data/external-tool/ExternalToolLaunchState.composable.ts
@@ -1,5 +1,5 @@
 import { useExternalToolApi } from "./ExternalToolApi.composable";
-import { ToolLaunchRequest, ToolLaunchRequestMethodEnum } from "@/store/external-tool";
+import { ToolLaunchRequest, ToolLaunchRequestMethodEnum } from "./types";
 import { BusinessError } from "@/store/types/commons";
 import { HttpStatusCode } from "@/store/types/http-status-code.enum";
 import { mapAxiosErrorToResponseError } from "@/utils/api";
