+++ src/modules/data/external-tool/SchoolExternalToolApi.composable.unit.ts
@@ -1,5 +1,5 @@
 import { useSchoolExternalToolApi } from "./SchoolExternalToolApi.composable";
-import { SchoolExternalToolMetadata } from "@/store/external-tool";
+import { SchoolExternalToolMetadata } from "./types";
 import { mockApi, mockApiResponse, schoolExternalToolMetadataResponseFactory } from "@@/tests/test-utils";
 import { SchoolExternalToolMetadataResponse } from "@api-server";
 import * as serverApi from "@api-server";
