+++ src/modules/data/external-tool/SchoolExternalToolUsage.composable.unit.ts
@@ -1,6 +1,6 @@
 import { useSchoolExternalToolApi } from "./SchoolExternalToolApi.composable";
 import { useSchoolExternalToolUsage } from "./SchoolExternalToolUsage.composable";
-import { SchoolExternalToolMetadata } from "@/store/external-tool";
+import { SchoolExternalToolMetadata } from "./types";
 import { BusinessError } from "@/store/types/commons";
 import { mapAxiosErrorToResponseError } from "@/utils/api";
 import { axiosErrorFactory, mockComposable, schoolExternalToolMetadataFactory } from "@@/tests/test-utils";
