+++ src/modules/data/external-tool/school-external-tools.composable.unit.ts
@@ -1,4 +1,4 @@
-import { SchoolExternalTool } from "../../../store/external-tool";
+import { SchoolExternalTool } from "./types";
 import {
 	expectNotification,
 	mockApi,
