--- src/modules/data/external-tool/types/context-external-tool-configuration-status.ts
@@ -1,13 +1,0 @@
-export interface ContextExternalToolConfigurationStatus {
-	isOutdatedOnScopeSchool: boolean;
-
-	isOutdatedOnScopeContext: boolean;
-
-	isIncompleteOnScopeContext: boolean;
-
-	isIncompleteOperationalOnScopeContext: boolean;
-
-	isDeactivated: boolean;
-
-	isNotLicensed: boolean;
-}
