--- src/modules/data/external-tool/types/context-external-tool-template-list-item.ts
@@ -1,6 +1,0 @@
-export interface ContextExternalToolTemplateListItem {
-	id: string;
-	name: string;
-	logoUrl?: string;
-	schoolToolId: string;
-}
