+++ src/modules/data/external-tool/types/external-tool-display-data.ts
@@ -1,24 +1,15 @@
-import { ContextExternalToolConfigurationStatus } from "./context-external-tool-configuration-status";
+import { ContextExternalToolConfigurationStatus } from "./context-external-tool";
 import { LtiDeepLinkResponse } from "@api-server";
 
 export type ExternalToolDisplayData = {
 	contextExternalToolId: string;
-
 	name: string;
-
 	domain: string;
-
 	description?: string;
-
 	logoUrl?: string;
-
 	thumbnailUrl?: string;
-
 	openInNewTab: boolean;
-
 	status: ContextExternalToolConfigurationStatus;
-
 	isLtiDeepLinkingTool: boolean;
-
 	ltiDeepLink?: LtiDeepLinkResponse;
 };
