+++ src/modules/data/external-tool/types/index.ts
@@ -1,5 +1,8 @@
+export * from "../types/external-tool.mapper";
+export * from "../types/school-external-tool.mapper";
 export * from "./context-external-tool";
-export * from "./context-external-tool-configuration-status";
-export * from "./context-external-tool-template-list-item";
 export * from "./external-tool-display-data";
+export * from "./school-external-tool";
 export * from "./tool-configuration-template";
+export * from "./tool-launch-request";
+export * from "./tool-parameter";
