+++ src/modules/data/external-tool/types/tool-launch-request.ts
@@ -1,5 +1,9 @@
-import { ToolLaunchRequestMethodEnum } from "@/store/external-tool/tool-launch-request-method.enum";
 import { LaunchRequestMethod, LaunchType } from "@api-server";
+
+export enum ToolLaunchRequestMethodEnum {
+	Get = "GET",
+	Post = "POST",
+}
 
 export type ToolLaunchRequest = {
 	method: ToolLaunchRequestMethodEnum | LaunchRequestMethod;
