--- src/store/external-tool/index.ts
@@ -1,10 +1,0 @@
-export * from "./school-external-tool";
-export * from "./school-external-tool-configuration-status";
-export * from "./school-external-tool-metadata";
-export * from "./tool-launch-request";
-export * from "./tool-launch-request-method.enum";
-export * from "./tool-parameter";
-export * from "./tool-parameter.enum";
-export * from "./tool-parameter-entry";
-export * from "./tool-parameter-location.enum";
-export * from "./tool-parameter-scope.enum";
