--- src/store/external-tool/mapper/index.ts
@@ -1,2 +1,0 @@
-export * from "./external-tool.mapper";
-export * from "./school-external-tool.mapper";
