--- src/store/external-tool/school-external-tool-configuration-status.ts
@@ -1,4 +1,0 @@
-export interface SchoolExternalToolConfigurationStatus {
-	isOutdatedOnScopeSchool: boolean;
-	isGloballyDeactivated: boolean;
-}
