--- src/store/external-tool/school-external-tool-metadata.ts
@@ -1,5 +1,0 @@
-export interface SchoolExternalToolMetadata {
-	course: number;
-	boardElement: number;
-	mediaBoard: number;
-}
