--- src/store/external-tool/tool-launch-request-method.enum.ts
@@ -1,4 +1,0 @@
-export enum ToolLaunchRequestMethodEnum {
-	Get = "GET",
-	Post = "POST",
-}
