--- src/store/external-tool/tool-parameter-entry.ts
@@ -1,5 +1,0 @@
-export interface ToolParameterEntry {
-	name: string;
-
-	value?: string;
-}
