--- src/store/external-tool/tool-parameter-location.enum.ts
@@ -1,6 +1,0 @@
-export enum ToolParameterLocation {
-	PATH = "path",
-	BODY = "body",
-	QUERY = "query",
-	FRAGMENT = "fragment",
-}
