--- src/store/external-tool/tool-parameter-scope.enum.ts
@@ -1,5 +1,0 @@
-export enum ToolParameterScope {
-	Global = "global",
-	School = "school",
-	Context = "context",
-}
