--- src/store/external-tool/tool-parameter.enum.ts
@@ -1,12 +1,0 @@
-export enum ToolParameterType {
-	String = "string",
-	Number = "number",
-	Boolean = "boolean",
-	AutoContextid = "auto_contextid",
-	AutoContextname = "auto_contextname",
-	AutoSchoolid = "auto_schoolid",
-	AutoSchoolnumber = "auto_schoolnumber",
-	AutoMediumid = "auto_mediumid",
-	AutoGroupExternaluuid = "auto_group_externaluuid",
-	AutoPublisher = "auto_publisher",
-}
