+++ tests/test-utils/factory/toolParameterEntryFactory.ts
@@ -1,4 +1,4 @@
-import { ToolParameterEntry } from "@/store/external-tool";
+import { ToolParameterEntry } from "@data-external-tool";
 import { Factory } from "fishery";
 
 export const toolParameterEntryFactory = Factory.define<ToolParameterEntry>(({ sequence }) => ({
